Update Q6, 7
Q6 correctly calculates the original value and removes the sale quantity from both itself and the targeted block.

Q7 asks for company and stores multiple queues in a hashmap. if an inputted company matches a key stock will be sold for it, otherwise you are told the company does not exist.