Update Q8
Added methods to get sum product difference and quotients. The order of operations is backwards from the order arguments are inputted to accomodate how numbers are placed on the stack.

A switch determines what operation should be carried out when evaluating the top

TODO work out the criteria for when to evaluate top. The top may be evaluated differently depending on whether there are open parentheses or multiplication symbols still present...