Update Q9, Q10
Q9---

Added Path class containing coordinates and direction of each path so that they may be added to a stack for backtracking.

Changed the int[][] to a char array using boxes and circles to improve the mazes appearance.

traverse returns coordinates 1 above from a given direction.

getChar returns the character at a set of coordinates

IsIntersection checks if any paths intersect the current path, ignoring directions of the same axis.

pushPossiblePaths checks every direction for potential paths. If a direction is specified it will ignore any direction in each axix, this is useful for intersections.

The X and Y in display had to be inverted due to how the char array was created. This means that coordinates are given in (y, x) as opposed to (x, y)

TODO detect when exit has been reached. This may be done by returning a special character from getChar

Q10---
Added cities to map.txt
The strings will be imported bidirectionally, meaning each connection only needs to be written once.